backend/requirements.txt
# FastAPI and Core Dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=42.0.0

# Database & ORM
supabase==2.4.2
psycopg2-binary==2.9.9

# GitHub Integration
PyGithub==2.1.1
GitPython==3.1.42

# HTTP Client
httpx==0.27.2
requests==2.31.0

# Validation & Serialization
pydantic[email]==2.12.0
pydantic-settings==2.2.1

# Utility
redis[hiredis]==5.0.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
isort==5.13.2
mypy==1.8.0

# File Operations
aiofiles==23.2.1
PyYAML==6.0.1

# ============================================
# AI & NLP Dependencies
# ============================================

# NLP & Text Processing
spacy==3.7.2
transformers==4.37.2

# LLM API Clients
groq>=0.4.0
google-generativeai>=0.4.0
openai>=1.12.0

# Ethics & Bias Detection
aif360==0.5.0

# WebSocket Support
websockets==12.0

# Additional Utilities
tenacity>=8.0.0